feat: Rework example and relocate controllers
