+++ example/src/main/resources/io/github/sekassel/jfxexample/view/back.fxml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.VBox?>
+<VBox alignment="CENTER" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity"
+      xmlns="http://javafx.com/javafx/19"
+      xmlns:fx="http://javafx.com/fxml/1" fx:controller="io.github.sekassel.jfxexample.controller.BackController">
+    <Button fx:id="back" onAction="#back" text="Back"/>
+</VBox>
