+++ example/src/main/resources/io/github/sekassel/jfxexample/view/main.fxml
@@ -6,11 +6,11 @@
 <?import javafx.scene.control.Button?>
 <VBox alignment="CENTER" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity"
       prefHeight="400.0" prefWidth="600.0" spacing="20.0" xmlns="http://javafx.com/javafx/19"
-      xmlns:fx="http://javafx.com/fxml/1" fx:controller="io.github.sekassel.jfxexample.controller2.MainController">
+      xmlns:fx="http://javafx.com/fxml/1" fx:controller="io.github.sekassel.jfxexample.controller.MainController">
     <padding>
         <Insets bottom="30.0" left="30.0" right="30.0" top="30.0"/>
     </padding>
     <Label/>
     <Label fx:id="welcomeLabel" layoutX="310.0" layoutY="202.0"/>
-    <Button fx:id="back" onAction="#back" text="Back"/>
+    <Button fx:id="continueButton" onAction="#continueButton" text="Continue"/>
 </VBox>
